% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeppelin.R
\docType{package}
\name{zeppelin}
\alias{zeppelin}
\alias{zeppelin-package}
\alias{zeppelin-package}
\title{Build jekyll sites}
\description{
Site builder for https://github.com/nteetor/yonder.
}
\seealso{
For more on Jekyll check out \url{https://jekyllrb.com/docs/home/}.

For more on using Jekyll and Github pages check out,
\url{https://jekyllrb.com/docs/github-pages/}.
}
\author{
\strong{Maintainer}: Nathan Teetor \email{nathanteetor@gmail.com}

}
